+++ libs/designer-ui/src/lib/expressioneditor/index.tsx
@@ -1,3 +1,4 @@
+import type { EditorContentChangedEventArgs } from '../editor/monaco';
 import { MonacoEditor } from '../editor/monaco';
 import type { EventHandler } from '../eventhandler';
 import { EditorLanguage, clamp } from '@microsoft/logic-apps-shared';
@@ -22,6 +23,9 @@ export interface ExpressionEditorProps {
   onBlur?: EventHandler<ExpressionEditorEvent>;
   setIsDragging: (isDragging: boolean) => void;
   setExpressionEditorError: (error: string) => void;
+  onFocus?: () => void;
+  onContentChanged?(e: EditorContentChangedEventArgs): void;
+  isReadOnly?: boolean;
 }
 
 export function ExpressionEditor({
@@ -33,8 +37,11 @@ export function ExpressionEditor({
   currentHeight,
   setCurrentHeight,
   onBlur,
+  onFocus,
   setIsDragging,
   setExpressionEditorError,
+  onContentChanged,
+  isReadOnly = false,
 }: ExpressionEditorProps): JSX.Element {
   const [mouseDownLocation, setMouseDownLocation] = useState(0);
   const [heightOnMouseDown, setHeightOnMouseDown] = useState(0);
@@ -76,13 +83,15 @@ export function ExpressionEditor({
         overviewRulerBorder={false}
         contextMenu={false}
         onBlur={handleBlur}
-        onContentChanged={handleChangeEvent}
+        onFocus={onFocus}
+        onContentChanged={onContentChanged ?? handleChangeEvent}
         width={'100%'}
         wordWrap="bounded"
         wordWrapColumn={200}
         automaticLayout={true}
         data-automation-id="msla-expression-editor"
         height={`${currentHeight}px`}
+        readOnly={isReadOnly}
       />
       <div
         className="msla-expression-editor-expand"